[VYMCA-48] Small refactoring
