+++ js/gated-content/src/components/event/LiveStreamListing.vue
@@ -37,9 +37,9 @@
 
 <script>
 import client from '@/client';
-import LiveStreamTeaser from '@/components/LiveStreamTeaser.vue';
+import LiveStreamTeaser from '@/components/event/LiveStreamTeaser.vue';
 import Spinner from '@/components/Spinner.vue';
-import { JsonApiCombineMixin } from '../mixins/JsonApiCombineMixin';
+import { JsonApiCombineMixin } from '@/mixins/JsonApiCombineMixin';
 
 export default {
   name: 'LiveStreamListing',
