+++ js/gated-content/src/components/video/CategoryTeaser.vue
