+++ js/gated-content/src/views/CategoryPage.vue
@@ -22,8 +22,8 @@
 <script>
 import client from '@/client';
 import 'vue-lazy-youtube-video/dist/style.css';
-import VideoListing from '../components/VideoListing.vue';
-import { JsonApiCombineMixin } from '../mixins/JsonApiCombineMixin';
+import VideoListing from '@/components/video/VideoListing.vue';
+import { JsonApiCombineMixin } from '@/mixins/JsonApiCombineMixin';
 
 export default {
   name: 'CategoryPage',
